feat(wysiwyg): Parse HTML fragments by using `DOMParser`.
This patch adds the basics for parsing HTML fragments with `DOMParser` instead
of `html5ever` when the `js` feature is turned on. Why? Because `html5ever`
is really heavy. The idea is to use the [`DOMParser.parseFromString`]
(https://developer.mozilla.org/en-US/docs/Web/ API/ DOMParser/parseFromString)
function to parse HTML fragments and to rebuilt a `Dom` tree from the returned
`HTMLDocument`.

The immediate impacts is that `wysiwyg-wasm`, once compiled to a Wasm module,
drops from 522Kb to 123Kb. That's a size reduction of 76%.

Ideally, I would like to re-use the existing `parser::parse` tests, but
`speculoos` uses `num`, which uses `rustc-serialize`, which doesn't compile to
Wasm properly.

To run the tests, simply run:

```sh
$ wasm-pack test --release --firefox --headless -- --no-default-features --features js -- parser::parse
```

It is not possible to run the tests with NodeJS because this runtime doesn't
provide the `DOMParser` API.