feat: Add `insert_parent` DOM method (#415)
Adds a DOM method to insert a parent node around a range.

Currently, when adding a hyperlink to a range, it is done by applying the link to every leaf in the range. In order to wrap the leaves at a higher level in the tree, it'd be useful to have a method at the Dom level to insert a parent node around a range.