Add transaction functionality to Dom
Adds the ability to perform a set of operations in a transactional
manner on the Dom class. The transaction is indicated by the new member
variable 'is_transaction_in_progess', which is set to 'true' when the
operations start and 'false' when they complete. The class now includes
methods for starting and ending a transaction, which enforce class
invariants and ensure that only one transaction can be active at a time.