[Android] Fix performance issue in recomposition
The composable's `update` block may be called frequently regardless
of whether the `style` argument has changed. And setting the style of
the editor forces a re-render whether or not the style has actually
changed.

This change ensures that the style update and subsequent re-render is
only triggered if the new style is different to the old style.